Fix #23: Fixed another an error that would occur if the Zend OPcache extension was installed, but its "restrict_api" setting was enabled
If the OPcache is set to restrictive, it will throw an exception on this line and abort the entire install process. As this line is already checked if it exists or not, it does not seem to be crucial for the installation process. The loadExtensions function has the same check and already throws away any errors as an example.